# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Ms the trend API",
        "description": "API documentation for the Ms the trend project.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/order-save": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Create a new order",
                "operationId": "60877667d6f133e309b2ae6e149520ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "product_id": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2
                                        ]
                                    },
                                    "quantity": {
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            2,
                                            1
                                        ]
                                    },
                                    "price": {
                                        "type": "array",
                                        "items": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "example": [
                                            500,
                                            1200
                                        ]
                                    },
                                    "amount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": 1700
                                    },
                                    "payment_id": {
                                        "type": "string",
                                        "example": "pay_12345"
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "123 Main St, NY"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/update-customer-address": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update customer address for a single order detail",
                "operationId": "440ecaebe4ac16f5c1bee4a7ef11b36b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "New Street 456, LA"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer address updated successfully"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Order detail not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cancel-order/{id}": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Cancel an order",
                "operationId": "7e0eab4aeb1976da38d1c20512afce2d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrderDetail ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order cancelled successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/addAddress": {
            "post": {
                "tags": [
                    "Orders"
                ],
                "summary": "Update customer address for all order details in an order",
                "operationId": "2a4ef678f2c484e281a4a52d5f31e34f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "Order ID",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "customer_address": {
                                        "type": "string",
                                        "example": "45 Downtown Street, Chicago"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Customer address updated successfully for all order items"
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}